+++ modal_version/_version_generated.py
@@ -1,4 +1,4 @@
 # Copyright Modal Labs 2025
 
 # Note: Reset this value to -1 whenever you make a minor `0.X` release of the client.
-build_number = 165  # git: 063c3c2
+build_number = 5  # git: 2e65784